requirements.txt
fastapi>=0.95.0
uvicorn>=0.23.1
python-telegram-bot>=20.0
langchain>=0.0.300
faiss-cpu>=1.7.4
openai>=0.27.0
python-dotenv>=1.0.0
